make abstract data module
